Refactored user data
