+++ src/user/models/ident/dto/ident-user-data.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { IsNotEmpty, IsEnum, IsString, IsObject, ValidateIf } from 'class-validator';
 import { Country } from 'src/shared/models/country/country.entity';
-import { AccountType } from '../../userData/account-type.enum';
+import { AccountType } from '../../userData/userData.entity';
 
 export class IdentUserDataDto {
   @ApiProperty()
