+++ src/user/models/limit-request/limit-request.service.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { KycDocument } from 'src/user/services/kyc/dto/kyc.dto';
 import { KycService } from 'src/user/services/kyc/kyc.service';
-import { kycCompleted } from '../userData/userData.entity';
+import { KycCompleted } from '../userData/userData.entity';
 import { UserDataService } from '../userData/userData.service';
 import { LimitRequestDto } from './dto/limit-request.dto';
 import { LimitRequest } from './limit-request.entity';
@@ -18,7 +18,7 @@ export class LimitRequestService {
   async increaseLimit(userId: number, dto: LimitRequestDto): Promise<LimitRequest> {
     // get user data
     const user = await this.userDataService.getUserDataForUser(userId);
-    if (!kycCompleted(user.kycStatus)) throw new BadRequestException('KYC not yet completed');
+    if (!KycCompleted(user.kycStatus)) throw new BadRequestException('KYC not yet completed');
 
     // create entity
     const entity = this.limitRequestRepo.create(dto);
