+++ src/user/models/user/user.entity.ts
@@ -8,7 +8,7 @@ import { Wallet } from 'src/user/models/wallet/wallet.entity';
 import { Entity, Column, OneToMany, ManyToOne } from 'typeorm';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { Fiat } from 'src/shared/models/fiat/fiat.entity';
-import { AccountType } from '../userData/account-type.enum';
+import { AccountType } from '../userData/userData.entity';
 import { Staking } from '../../../payment/models/staking/staking.entity';
 import { IEntity } from 'src/shared/models/entity';
