--- src/user/models/userData/account-type.enum.ts
@@ -1,5 +1,0 @@
-export enum AccountType {
-  PERSONAL = 'Personal',
-  BUSINESS = 'Business',
-  SOLE_PROPRIETORSHIP = 'SoleProprietorship',
-}
