+++ src/user/models/userData/userData.repository.ts
@@ -2,8 +2,4 @@ import { EntityRepository, Repository } from 'typeorm';
 import { UserData } from './userData.entity';
 
 @EntityRepository(UserData)
-export class UserDataRepository extends Repository<UserData> {
-  async getAllUserData(): Promise<UserData[]> {
-    return await this.find({ relations: ['bankDatas', 'users'] });
-  }
-}
+export class UserDataRepository extends Repository<UserData> {}
